chore: moves policies under models and comments out AppServiceProvider leaving it to be auto-discovered; minor models code adjustments like adds missing types, etc.
