+++ app/Models/CompetenceTranslation.php
@@ -6,11 +6,15 @@ namespace App\Models;
 
 use App\Concerns\HasSlug;
 use App\Enums\CompetenceStatus;
+use Database\Factories\CompetenceTranslationFactory;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Carbon;
 
 /**
  * @property string $name
@@ -18,27 +22,27 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * @property string $excerpt
  * @property string $body
  * @property int $order
- * @property \App\Enums\CompetenceStatus $status
- * @property \Illuminate\Support\Carbon $published_at
- * @property \Illuminate\Support\Carbon $published_through
- * @property \Illuminate\Support\Carbon $created_at
- * @property \Illuminate\Support\Carbon $updated_at
- * @property \Illuminate\Support\Carbon $deleted_at
+ * @property CompetenceStatus $status
+ * @property Carbon $published_at
+ * @property Carbon $published_through
+ * @property Carbon $created_at
+ * @property Carbon $updated_at
+ * @property Carbon $deleted_at
  * @property int $competence_id
  * @property int $language_id
  * @property int $user_id
- * @property-read \App\Models\Competence $competence
- * @property-read \App\Models\Language $language
- * @property-read \App\Models\User $user
- * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Tag> $tags
+ * @property-read Competence $competence
+ * @property-read Language $language
+ * @property-read User $user
+ * @property-read Collection<int, Tag> $tags
  *
- * @mixin \Illuminate\Database\Eloquent\Builder
+ * @mixin Builder
  *
- * @extends \Illuminate\Database\Eloquent\Model<\Database\Factories\CompetenceTranslationFactory>
+ * @extends Model<CompetenceTranslationFactory>
  */
 class CompetenceTranslation extends Model
 {
-    /** @use HasFactory<\Database\Factories\CompetenceTranslationFactory> */
+    /** @use HasFactory<CompetenceTranslationFactory> */
     use HasFactory, HasSlug, SoftDeletes;
 
     /**
@@ -47,18 +51,6 @@ class CompetenceTranslation extends Model
      * @var array<string>
      */
     protected $guarded = ['id'];
-
-    protected $casts = [
-        'excerpt' => 'string',
-        'body' => 'string',
-        'order' => 'integer',
-        'status' => CompetenceStatus::class,
-        'published_at' => 'datetime',
-        'published_through' => 'datetime',
-        'created_at' => 'datetime',
-        'updated_at' => 'datetime',
-        'deleted_at' => 'datetime',
-    ];
 
     /**
      * Get the attributes that should be cast.
@@ -89,7 +81,7 @@ class CompetenceTranslation extends Model
     }
 
     /**
-     * Get all of the tags for the CompetenceTranslation
+     * Get all the tags for the CompetenceTranslation
      */
     public function tags(): MorphToMany
     {
