+++ app/Models/Gallery.php
@@ -6,26 +6,30 @@ namespace App\Models;
 
 use App\Concerns\HasSlug;
 use Database\Factories\GalleryFactory;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Carbon;
 
 /**
  * @property string $name
  * @property string $slug
- * @property \Illuminate\Support\Carbon $created_at
- * @property \Illuminate\Support\Carbon $updated_at
- * @property \Illuminate\Support\Carbon $deleted_at
+ * @property Carbon $created_at
+ * @property Carbon $updated_at
+ * @property Carbon $deleted_at
  * @property int $gallery_id
- * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Tag> $tags
- * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Image> $images
- * @property-read \App\Models\Gallery $galleries
+ * @property-read Collection<int, Tag> $tags
+ * @property-read Collection<int, Image> $images
+ * @property-read Gallery $parentGallery
  *
- * @mixin \Illuminate\Database\Eloquent\Builder
+ * @mixin Builder
  *
- * @extends \Illuminate\Database\Eloquent\Model<\Database\Factories\GalleryFactory>
+ * @extends Model<GalleryFactory>
  */
 class Gallery extends Model
 {
@@ -40,15 +44,15 @@ class Gallery extends Model
     protected $guarded = ['id'];
 
     /**
-     * Get all of the tags for the Gallery
+     * Get all the tags for the Gallery
      */
-    public function tags(): \Illuminate\Database\Eloquent\Relations\MorphToMany
+    public function tags(): MorphToMany
     {
         return $this->morphToMany(Tag::class, 'taggable');
     }
 
     /**
-     * Get all of the images for the Gallery
+     * Get all the images for the Gallery
      */
     public function images(): HasMany
     {
@@ -58,7 +62,7 @@ class Gallery extends Model
     /**
      * Get the galleries that owns the Gallery
      */
-    public function galleries(): BelongsTo
+    public function parentGallery(): BelongsTo
     {
         return $this->belongsTo(Gallery::class, 'gallery_id', 'id');
     }
