+++ app/Models/Policies/CustomerPolicy.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Policies;
+namespace App\Models\Policies;
 
-use App\Models\Menu;
+use App\Models\Customer;
 use App\Models\User;
 
-class MenuPolicy
+class CustomerPolicy
 {
     /**
      * Determine whether the user can view any models.
@@ -18,7 +18,7 @@ class MenuPolicy
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Menu $menu): bool
+    public function view(User $user, Customer $customer): bool
     {
         return false;
     }
@@ -34,7 +34,7 @@ class MenuPolicy
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Menu $menu): bool
+    public function update(User $user, Customer $customer): bool
     {
         return false;
     }
@@ -42,7 +42,7 @@ class MenuPolicy
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Menu $menu): bool
+    public function delete(User $user, Customer $customer): bool
     {
         return false;
     }
@@ -50,7 +50,7 @@ class MenuPolicy
     /**
      * Determine whether the user can restore the model.
      */
-    public function restore(User $user, Menu $menu): bool
+    public function restore(User $user, Customer $customer): bool
     {
         return false;
     }
@@ -58,7 +58,7 @@ class MenuPolicy
     /**
      * Determine whether the user can permanently delete the model.
      */
-    public function forceDelete(User $user, Menu $menu): bool
+    public function forceDelete(User $user, Customer $customer): bool
     {
         return false;
     }
