+++ app/Models/Policies/PostPolicy.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Policies;
+namespace App\Models\Policies;
 
-use App\Models\Translation;
+use App\Models\Post;
 use App\Models\User;
 
-class TranslationPolicy
+class PostPolicy
 {
     /**
      * Determine whether the user can view any models.
@@ -18,7 +18,7 @@ class TranslationPolicy
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Translation $translation): bool
+    public function view(User $user, Post $post): bool
     {
         return false;
     }
@@ -34,7 +34,7 @@ class TranslationPolicy
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Translation $translation): bool
+    public function update(User $user, Post $post): bool
     {
         return false;
     }
@@ -42,7 +42,7 @@ class TranslationPolicy
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Translation $translation): bool
+    public function delete(User $user, Post $post): bool
     {
         return false;
     }
@@ -50,7 +50,7 @@ class TranslationPolicy
     /**
      * Determine whether the user can restore the model.
      */
-    public function restore(User $user, Translation $translation): bool
+    public function restore(User $user, Post $post): bool
     {
         return false;
     }
@@ -58,7 +58,7 @@ class TranslationPolicy
     /**
      * Determine whether the user can permanently delete the model.
      */
-    public function forceDelete(User $user, Translation $translation): bool
+    public function forceDelete(User $user, Post $post): bool
     {
         return false;
     }
