+++ app/Models/Policies/TranslatePolicy.php
@@ -1,23 +1,24 @@
 <?php
 
-namespace App\Policies;
+namespace App\Models\Policies;
 
+use App\Models\Translate;
 use App\Models\User;
 
-class UserPolicy
+class TranslatePolicy
 {
     /**
      * Determine whether the user can view any models.
      */
     public function viewAny(User $user): bool
     {
-        return true;
+        return false;
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, User $model): bool
+    public function view(User $user, Translate $translate): bool
     {
         return false;
     }
@@ -33,7 +34,7 @@ class UserPolicy
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, User $model): bool
+    public function update(User $user, Translate $translate): bool
     {
         return false;
     }
@@ -41,7 +42,7 @@ class UserPolicy
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, User $model): bool
+    public function delete(User $user, Translate $translate): bool
     {
         return false;
     }
@@ -49,7 +50,7 @@ class UserPolicy
     /**
      * Determine whether the user can restore the model.
      */
-    public function restore(User $user, User $model): bool
+    public function restore(User $user, Translate $translate): bool
     {
         return false;
     }
@@ -57,8 +58,8 @@ class UserPolicy
     /**
      * Determine whether the user can permanently delete the model.
      */
-    public function forceDelete(User $user, User $model): bool
+    public function forceDelete(User $user, Translate $translate): bool
     {
-        return true;
+        return false;
     }
 }
