+++ app/Models/Policies/UserPolicy.php
@@ -1,24 +1,23 @@
 <?php
 
-namespace App\Policies;
+namespace App\Models\Policies;
 
-use App\Models\Post;
 use App\Models\User;
 
-class PostPolicy
+class UserPolicy
 {
     /**
      * Determine whether the user can view any models.
      */
     public function viewAny(User $user): bool
     {
-        return false;
+        return true;
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Post $post): bool
+    public function view(User $user, User $model): bool
     {
         return false;
     }
@@ -34,7 +33,7 @@ class PostPolicy
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Post $post): bool
+    public function update(User $user, User $model): bool
     {
         return false;
     }
@@ -42,7 +41,7 @@ class PostPolicy
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Post $post): bool
+    public function delete(User $user, User $model): bool
     {
         return false;
     }
@@ -50,7 +49,7 @@ class PostPolicy
     /**
      * Determine whether the user can restore the model.
      */
-    public function restore(User $user, Post $post): bool
+    public function restore(User $user, User $model): bool
     {
         return false;
     }
@@ -58,8 +57,8 @@ class PostPolicy
     /**
      * Determine whether the user can permanently delete the model.
      */
-    public function forceDelete(User $user, Post $post): bool
+    public function forceDelete(User $user, User $model): bool
     {
-        return false;
+        return true;
     }
 }
