+++ app/Models/Policies/WorkPolicy.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Policies;
+namespace App\Models\Policies;
 
-use App\Models\Image;
 use App\Models\User;
+use App\Models\Work;
 
-class ImagePolicy
+class WorkPolicy
 {
     /**
      * Determine whether the user can view any models.
@@ -18,7 +18,7 @@ class ImagePolicy
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Image $photo): bool
+    public function view(User $user, Work $work): bool
     {
         return false;
     }
@@ -34,7 +34,7 @@ class ImagePolicy
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Image $photo): bool
+    public function update(User $user, Work $work): bool
     {
         return false;
     }
@@ -42,7 +42,7 @@ class ImagePolicy
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Image $photo): bool
+    public function delete(User $user, Work $work): bool
     {
         return false;
     }
@@ -50,7 +50,7 @@ class ImagePolicy
     /**
      * Determine whether the user can restore the model.
      */
-    public function restore(User $user, Image $photo): bool
+    public function restore(User $user, Work $work): bool
     {
         return false;
     }
@@ -58,7 +58,7 @@ class ImagePolicy
     /**
      * Determine whether the user can permanently delete the model.
      */
-    public function forceDelete(User $user, Image $photo): bool
+    public function forceDelete(User $user, Work $work): bool
     {
         return false;
     }
