+++ app/Models/PostTranslation.php
@@ -7,11 +7,14 @@ namespace App\Models;
 use App\Concerns\HasSlug;
 use App\Enums\PostStatus;
 use Database\Factories\PostTranslationFactory;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Carbon;
 
 /**
  * @property string $name
@@ -19,23 +22,23 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * @property string $excerpt
  * @property string $body
  * @property int $order
- * @property \App\Enums\PostStatus $status
- * @property \Illuminate\Support\Carbon $published_at
- * @property \Illuminate\Support\Carbon $published_through
- * @property \Illuminate\Support\Carbon $created_at
- * @property \Illuminate\Support\Carbon $updated_at
- * @property \Illuminate\Support\Carbon $deleted_at
+ * @property PostStatus $status
+ * @property Carbon $published_at
+ * @property Carbon $published_through
+ * @property Carbon $created_at
+ * @property Carbon $updated_at
+ * @property Carbon $deleted_at
  * @property int $post_id
  * @property int $language_id
  * @property int $user_id
- * @property-read \App\Models\Post $post
- * @property-read \App\Models\Language $language
- * @property-read \App\Models\User $user
- * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Tag> $tags
+ * @property-read Post $post
+ * @property-read Language $language
+ * @property-read User $user
+ * @property-read Collection<int, Tag> $tags
  *
- * @mixin \Illuminate\Database\Eloquent\Builder
+ * @mixin Builder
  *
- * @extends \Illuminate\Database\Eloquent\Model<\Database\Factories\PostTranslationFactory>
+ * @extends Model<PostTranslationFactory>
  */
 class PostTranslation extends Model
 {
@@ -78,7 +81,7 @@ class PostTranslation extends Model
     }
 
     /**
-     * Get all of the tags for the PostTranslation
+     * Get all the tags for the PostTranslation
      */
     public function tags(): MorphToMany
     {
