+++ app/Models/Translation.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace App\Models;
 
+use Database\Factories\TranslationFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Translation extends Model
 {
-    /** @use HasFactory<\Database\Factories\TranslationFactory> */
+    /** @use HasFactory<TranslationFactory> */
     use HasFactory;
 
     protected $table = 'translations';
