+++ bootstrap/providers.php
@@ -1,5 +1,5 @@
 <?php
 
 return [
-    App\Providers\AppServiceProvider::class,
+//    App\Providers\AppServiceProvider::class,
 ];
